+++ src/Authress.SDK/Client/HttpClientProvider.cs
@@ -32,15 +32,36 @@ namespace Authress.SDK.Client
         public TimeSpan? RequestTimeout { get; set; } = null;
     }
 
+    /// <summary>
+    /// Authress Domain Host: https://CUSTOM_DOMAIN.application.com (Get an authress custom domain: https://authress.io/app/#/settings?focus=domain)
+    /// </summary>
+    public class AuthressSettings
+    {
+        /// <summary>
+        /// Authress Domain Host: https://CUSTOM_DOMAIN.application.com (Get an authress custom domain: https://authress.io/app/#/settings?focus=domain)
+        /// </summary>
+        public string ApiBasePath { get; set; } = "https://api.authress.io";
+
+        /// <summary>
+        /// Timeout for requests to Authress. Default is unset.
+        /// </summary>
+        public TimeSpan? RequestTimeout { get; set; } = null;
+
+        /// <summary>
+        /// Max retries for requests that fail.
+        /// </summary>
+        public uint MaxRetries { get; set; } = 5;
+    }
+
     internal class HttpClientProvider
     {
         private readonly SemaphoreSlim syncObj = new SemaphoreSlim(1);
         private HttpClient clientProxy;
-        private readonly HttpClientSettings settings;
+        private readonly AuthressSettings settings;
         private readonly ITokenProvider tokenProvider;
         private readonly IHttpClientHandlerFactory customHttpClientHandlerFactory;
 
-        public HttpClientProvider(HttpClientSettings settings, ITokenProvider tokenProvider, IHttpClientHandlerFactory customHttpClientHandlerFactory = null)
+        public HttpClientProvider(AuthressSettings settings, ITokenProvider tokenProvider, IHttpClientHandlerFactory customHttpClientHandlerFactory = null)
         {
             this.settings = settings;
             this.tokenProvider = tokenProvider;
@@ -66,10 +87,26 @@ namespace Authress.SDK.Client
                 }
 
                 // create the inner handlers for the cache handler, the outermost handler is called first. (https://docs.microsoft.com/en-us/aspnet/web-api/overview/advanced/http-message-handlers)
+                // * The retry handler is listed first, then everything else will be executed once, and the retries will run in an inner loop
+                // * If the retry handler is listed last, then the whole stack is retried going all the way in and all the way back out to the retry handler.
+                // * If you want something to be recalculated on every retried call put it before the retry handler, if a handler is invariant across multiple calls, put it after the retry handler.
+                /**** ⌄ Called Last ⌄ ******/
                 HttpMessageHandler outermostHandler = customHttpClientHandlerFactory?.Create() ?? new HttpClientHandler { AllowAutoRedirect = true };
+                // List of Handlers that need to be recalculated on every call
+
+                // -
+                // -
+
+                /********************/
+                // The retry handler
+                outermostHandler = new RetryHandler(outermostHandler, settings.MaxRetries);
+                /********************/
+
+                // List of Handlers that never need to be retried
                 outermostHandler = new RewriteBaseUrlHandler(outermostHandler, settings.ApiBasePath);
                 outermostHandler = new AddAuthorizationHeaderHandler(outermostHandler, tokenProvider, settings.ApiBasePath);
                 outermostHandler = new AddUserAgentHeaderHandler(outermostHandler);
+                /**** ⌃ Called First ⌃ ******/
 
                 // create the client and assign it to the member variable for future access, create a tmp client so that the client is fully initialized before setting "client" property.
                 clientProxy = new HttpClient(outermostHandler);
@@ -90,8 +127,8 @@ namespace Authress.SDK.Client
 
     internal class AddAuthorizationHeaderHandler : DelegatingHandler
     {
-        private ITokenProvider tokenProvider;
-        private string apiBasePath;
+        private readonly ITokenProvider tokenProvider;
+        private readonly string apiBasePath;
 
         public AddAuthorizationHeaderHandler(HttpMessageHandler innerHandler, ITokenProvider tokenProvider, string apiBasePath) : base(innerHandler)
         {
@@ -101,9 +138,11 @@ namespace Authress.SDK.Client
 
         protected override async Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
         {
-            var bearerToken = await tokenProvider.GetBearerToken(apiBasePath);
-            var token = bearerToken.ToLower().Contains("bearer") ? bearerToken : $"Bearer {bearerToken}";
-            request.Headers.TryAddWithoutValidation("Authorization", token);
+            if (tokenProvider != null) {
+                var bearerToken = await tokenProvider.GetBearerToken(apiBasePath);
+                var token = bearerToken.ToLower().Contains("bearer") ? bearerToken : $"Bearer {bearerToken}";
+                request.Headers.TryAddWithoutValidation("Authorization", token);
+            }
             return await base.SendAsync(request, cancellationToken);
         }
     }
@@ -144,5 +183,51 @@ namespace Authress.SDK.Client
 
         private static string MergePath(string baseUrlPath, string requestPath) =>
             requestPath.StartsWith(baseUrlPath) ? requestPath : baseUrlPath + requestPath.Substring(1);
+    }
+
+    internal class RetryHandler : DelegatingHandler
+    {
+        private readonly uint maxRetries = 5;
+        private const int retryDelayMilliseconds = 20;
+
+        public RetryHandler(HttpMessageHandler innerHandler, uint maxRetries = 5) : base(innerHandler) {
+            this.maxRetries = maxRetries;
+        }
+
+        protected override async Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
+        {
+            HttpResponseMessage response = null;
+            Exception lastException = null;
+            for (int i = 0; i <= maxRetries; i++)
+            {
+                try
+                {
+                    response = await base.SendAsync(request, cancellationToken);
+
+                    if ((int) response.StatusCode <= 499)
+                    {
+                        return response;
+                    }
+                }
+                catch (Exception requestException)
+                {
+                    lastException = requestException;
+                }
+
+                if (i == maxRetries)
+                {
+                    break;
+                }
+
+                await Task.Delay((int) (retryDelayMilliseconds * Math.Pow(2, i)), cancellationToken);
+            }
+
+            if (response != null)
+            {
+                return response;
+            }
+
+            throw lastException;
+        }
     }
 }